Nexus/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
py-zabbix==1.1.7
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
swagger-client==1.0.0
urllib3==1.26.12
zmpe==1.0.1
